refactor(usecase-pattern): apply usecase pattern to weight and template search modules
