+++ src/modules/diet/macro-target/application/macroTarget.ts
@@ -7,7 +7,7 @@ import { type MacroProfile } from '~/modules/diet/macro-profile/domain/macroProf
 import { inForceMacroProfile } from '~/modules/diet/macro-profile/domain/macroProfileOperations'
 import { showError } from '~/modules/toast/application/toastManager'
 import { currentUserId } from '~/modules/user/application/user'
-import { userWeights } from '~/modules/weight/infrastructure/signals/weightsCacheStore'
+import { userWeights } from '~/modules/weight/application/usecases/weightState'
 import { inForceWeight } from '~/shared/utils/weightUtils'
 
 export const calculateMacroTarget = (
