+++ src/modules/weight/application/usecases/weightState.ts
@@ -1,9 +1,10 @@
 import { createResource } from 'solid-js'
 
 import { currentUserId } from '~/modules/user/application/user'
-import { createWeightCrudService } from '~/modules/weight/application/services/weightCrud'
+import { createWeightCrudService } from '~/modules/weight/application/usecases/weightCrud'
 import { weightSchema } from '~/modules/weight/domain/weight'
 import { createLocalStorageWeightRepository } from '~/modules/weight/infrastructure/localStorage/localStorageRepository'
+import { initializeWeightRealtime } from '~/modules/weight/infrastructure/supabase/realtime'
 import { createSupabaseWeightGateway } from '~/modules/weight/infrastructure/supabase/supabaseWeightGateway'
 import { createErrorHandler } from '~/shared/error/errorHandler'
 import { parseWithStack } from '~/shared/utils/parseWithStack'
@@ -12,20 +13,21 @@ const storageRepository = createLocalStorageWeightRepository()
 const weightRepository = createSupabaseWeightGateway()
 const errorHandler = createErrorHandler('application', 'Weight')
 
-// TODO: delete temp
-export const weightCrudService = createWeightCrudService({
-  weightRepository,
-  storageRepository,
-  errorHandler,
-  refetchUserWeights: () => void refetchUserWeights(),
-})
-
-export const [
+const [
   userWeights,
   { mutate: mutateUserWeights, refetch: refetchUserWeights },
 ] = createResource(
   currentUserId, // Source signal - refetches when userId changes
-  weightCrudService.fetchUserWeights,
+  async (userId: number) => {
+    try {
+      const weights = await weightRepository.fetchUserWeights(userId)
+      storageRepository.setCachedWeights(userId, weights)
+      return weights
+    } catch (error) {
+      errorHandler.error(error)
+      throw error
+    }
+  },
   {
     initialValue: parseWithStack(
       weightSchema.array(),
@@ -34,3 +36,16 @@ export const [
     ssrLoadFrom: 'initial',
   },
 )
+
+// CRUD operations service - temporary until fully migrated
+export const weightCrudService = createWeightCrudService({
+  weightRepository,
+  storageRepository,
+  errorHandler,
+})
+
+// Export state signals
+export { mutateUserWeights, refetchUserWeights, userWeights }
+
+// Initialize realtime subscription
+initializeWeightRealtime()
