+++ src/sections/profile/components/MacroEvolution.tsx
@@ -4,8 +4,8 @@ import { type DayDiet } from '~/modules/diet/day-diet/domain/dayDiet'
 import { type MacroProfile } from '~/modules/diet/macro-profile/domain/macroProfile'
 import { calculateMacroTarget } from '~/modules/diet/macro-target/application/macroTarget'
 import { CARD_BACKGROUND_COLOR, CARD_STYLE } from '~/modules/theme/constants'
+import { userWeights } from '~/modules/weight/application/usecases/weightState'
 import { type Weight } from '~/modules/weight/domain/weight'
-import { userWeights } from '~/modules/weight/infrastructure/signals/weightsCacheStore'
 import { dateToDDMM } from '~/shared/utils/date/dateUtils'
 import {
   calcCalories,
