+++ src/sections/search/components/TemplateSearchResults.tsx
@@ -11,7 +11,7 @@ import {
 import {
   debouncedTab,
   templates,
-} from '~/modules/template-search/application/templateSearch'
+} from '~/modules/template-search/application/usecases/templateSearchState'
 import { Alert } from '~/sections/common/components/Alert'
 import { RemoveFromRecentButton } from '~/sections/common/components/buttons/RemoveFromRecentButton'
 import { SearchLoadingIndicator } from '~/sections/search/components/SearchLoadingIndicator'
